+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * feat(ProseSteps): new component
 * feat(ProseSteps): new component
 * feat(ProseCard): new component
+* feat(ProseCardGroup): new component
 * feat(Banner): new component
 * feat(Card): new component
 * feat(CheckboxGroup): new component